requirements.txt
Flask==0.12.2
Flask-Cors==3.0.3
google-cloud-datastore==1.4.0
google-cloud-storage==1.11.0
google-cloud-logging==1.4.0
google-cloud-error_reporting==0.29.0
gunicorn==19.7.1
oauth2client==4.1.2
six==1.11.0
requests[security]==2.18.4
psq==0.6.0
